requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.34.0
anthropic==0.7.7
pydantic==2.5.0
alembic==1.13.0
pytest==7.4.3
httpx==0.25.2
